+++ scripts/pathfinding/pathfinding.js
@@ -4,6 +4,7 @@ CanvasQuadtree,
 ClockwiseSweepPolygon,
 CONFIG,
 CONST,
+foundry,
 PIXI
 */
 /* eslint no-unused-vars: ["error", { "argsIgnorePattern": "^_" }] */
@@ -403,37 +404,140 @@ export class Pathfinder {
  * @returns {PIXI.Point[]}
  */
 function cleanGridPath(pathPoints) {
-  const nPoints = pathPoints.length;
+  let nPoints = pathPoints.length;
   if ( nPoints < 3 ) return pathPoints;
+  // Debug: pathPoints.forEach(pt => Draw.point(pt, { alpha: 0.5, color: Draw.COLORS.blue }))
 
+  // const slowMethod = cleanGridPathSlow(pathPoints);
+
+  const orient2d = foundry.utils.orient2dFast;
   const config = { mode: "any", type: "move" };
+  let prev2;
   let prev = pathPoints[0];
   let curr = pathPoints[1];
-  const newPath = [prev];
+  let newPath = [prev];
   for ( let i = 2; i < nPoints; i += 1 ) {
     const next = pathPoints[i];
-    if ( next ) {
-      // If next shares this grid space, test if we can remove current.
-      const currGridPos = canvas.grid.grid.getGridPositionFromPixels(curr.x, curr.y);
-      const nextGridPos = canvas.grid.grid.getGridPositionFromPixels(next.x, next.y);
-      if ( currGridPos[0] === nextGridPos[0]
-        && currGridPos[1] === nextGridPos[1]
-        && !ClockwiseSweepPolygon.testCollision(prev, next, config) ) {
-        curr = next;
-        continue;
-      }
+
+    // Move points to the center of the grid square if no collision for previous or next.
+    const currCenter = getGridCenterPoint(curr);
+    if ( !(ClockwiseSweepPolygon.testCollision(prev, currCenter, config)
+      || ClockwiseSweepPolygon.testCollision(currCenter, next, config)) ) curr = currCenter;
+
+    // Remove duplicate points.
+    if ( curr.almostEqual(prev) ) {
+      curr = next;
+      continue;
     }
 
-    // Test if we can move the current point to the center of the grid without a collision.
-    // Don't overlap points, which can cause collisions.
-    const currCenter = getGridCenterPoint(curr);
-    if ( !prev.equals(currCenter) && !ClockwiseSweepPolygon.testCollision(prev, currCenter, config) ) curr = currCenter;
+    // Remove points in middle of straight line.
+    if ( prev2 && orient2d(prev2, prev, curr).almostEqual(0) ) newPath.pop();
+
     newPath.push(curr);
+    prev2 = prev;
     prev = curr;
     curr = next;
   }
-  newPath.push(pathPoints.at(-1));
+
+  // Remove point in middle of straight line at the end of the path.
+  const lastPoint = pathPoints.at(-1);
+  if ( newPath.length > 1
+    && orient2d(prev2, prev, lastPoint).almostEqual(0) ) newPath.pop();
+  newPath.push(lastPoint);
+
+
+  // Remove points in middle of straight line.
+//   nPoints = newPath.length;
+//   prev = newPath[0];
+//   curr = newPath[1];
+//   let filteredPath = [prev];
+//   for ( let i = 2; i < nPoints; i += 1 ) {
+//     const next = newPath[i];
+//     if ( orient2d(prev, curr, next).almostEqual(0) ) {
+//       curr = next;
+//       continue;
+//     }
+//     filteredPath.push(curr);
+//     prev = curr;
+//     curr = next;
+//   }
+//   filteredPath.push(newPath.at(-1));
+
+//   if ( slowMethod.length !== newPath.length ) console.debug("Slow Method returned different path", [...slowMethod], [...newPath]);
+//   for ( let i = 0; i < slowMethod.length; i += 1 ) {
+//     if ( !slowMethod[i].to2d().equals(newPath[i].to2d()) ) {
+//       console.debug("Slow Method returned different path", [...slowMethod], [...newPath]);
+//       break;
+//     }
+//   }
+
+
   return newPath;
+}
+
+
+/**
+ * For given point on a grid:
+ * - if next point shares this grid square, delete if prev --> next has no collision.
+ * - temporarily move to the grid center.
+ * - if collision, move back and go to next point. Otherwise keep at center.
+ * Don't move the start or end points.
+ * @param {PIXI.Point[]} pathPoints
+ * @returns {PIXI.Point[]}
+ */
+function cleanGridPathSlow(pathPoints) {
+  let nPoints = pathPoints.length;
+  if ( nPoints < 3 ) return pathPoints;
+  // Debug: pathPoints.forEach(pt => Draw.point(pt, { alpha: 0.5, color: Draw.COLORS.blue }))
+
+  // Move points to the center of the grid square if no collision for previous or next.
+  const config = { mode: "any", type: "move" };
+  let prev = pathPoints[0];
+  let curr = pathPoints[1];
+  let centeredPath = [prev];
+  for ( let i = 2; i < nPoints; i += 1 ) {
+    const next = pathPoints[i];
+    const currCenter = getGridCenterPoint(curr);
+    if ( !(ClockwiseSweepPolygon.testCollision(prev, currCenter, config)
+      || ClockwiseSweepPolygon.testCollision(currCenter, next, config)) ) curr = currCenter;
+    centeredPath.push(curr);
+    prev = curr;
+    curr = next;
+  }
+  centeredPath.push(pathPoints.at(-1));
+  // Debug: centeredPath.forEach(pt => Draw.point(pt, { alpha: 0.5, color: Draw.COLORS.green }))
+
+  // Remove duplicate points.
+  prev = centeredPath[0];
+  let dedupedPath = [prev];
+  for ( let i = 1; i < nPoints; i += 1 ) {
+    const curr = centeredPath[i];
+    if ( curr.almostEqual(prev) ) continue;
+    dedupedPath.push(curr);
+    prev = curr;
+  }
+  // Debug: dedupedPath.forEach(pt => Draw.point(pt, { color: Draw.COLORS.orange }))
+
+  // Remove points in middle of straight line.
+  const orient2d = foundry.utils.orient2dFast;
+  nPoints = dedupedPath.length;
+  prev = dedupedPath[0];
+  curr = dedupedPath[1];
+  let filteredPath = [prev];
+  for ( let i = 2; i < nPoints; i += 1 ) {
+    const next = dedupedPath[i];
+    if ( orient2d(prev, curr, next).almostEqual(0) ) {
+      curr = next;
+      continue;
+    }
+    filteredPath.push(curr);
+    prev = curr;
+    curr = next;
+  }
+  filteredPath.push(dedupedPath.at(-1));
+  // Debug: filteredPath.forEach(pt => Draw.point(pt))
+
+  return filteredPath;
 }
 
 function getGridCenterPoint(pt) {